% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders_ml.R
\name{build_outcome_formulas_min}
\alias{build_outcome_formulas_min}
\title{Build outcome nuisance formulas (minimal)}
\usage{
build_outcome_formulas_min(K, stems, rhs_builder = rhs_linear, ...)
}
\description{
Build outcome nuisance formulas (minimal)
}
