% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msnmm_pairs_methods.R
\name{coef.msnmm_pairs_fit}
\alias{coef.msnmm_pairs_fit}
\title{Coefficients for msnmm_pairs_fit}
\usage{
\method{coef}{msnmm_pairs_fit}(object, ...)
}
\arguments{
\item{object}{An msnmm_pairs_fit object.}
}
\description{
Coefficients for msnmm_pairs_fit
}
