% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\name{fit_glm}
\alias{fit_glm}
\title{Fit a GLM (wrapper)
Lightweight GLM wrapper that returns a predictor function}
\usage{
fit_glm(formula, data, family = gaussian(), ...)
}
\description{
Fit a GLM (wrapper)
Lightweight GLM wrapper that returns a predictor function
}
