% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msnmm_point.R
\name{msnmm_point}
\alias{msnmm_point}
\title{Point-exposure marginal SNMM with cross-fitting}
\usage{
msnmm_point(
  data,
  y0,
  y1,
  a,
  L0,
  Ltilde0,
  blip = ~1,
  q = "score",
  nuisance = list(pi = fit_glm(vars = L0, family = binomial()), mu = fit_glm(vars = L0,
    family = gaussian())),
  folds = 2,
  trim = 0,
  solver = c("nleqslv", "oneshot")[1],
  control = list(),
  seed = NULL
)
}
\arguments{
\item{data}{data.frame with columns y0,y1,a and specified covariates}

\item{y0, y1, a}{column names}

\item{L0}{adjustment set at baseline}

\item{Ltilde0}{effect-modifier set (for the blip)}

\item{blip}{formula for γ(ltilde; ψ), linear in ψ}

\item{q}{"score" (default) or a function(data)-> n x p matrix or a formula}

\item{nuisance}{list(pi=..., mu=...) each created with fit_glm/xgboost/SuperLearner}

\item{folds}{integer >=2}

\item{trim}{upper-trim on propensity (1 - π); 0 means none}
}
\description{
Point-exposure marginal SNMM with cross-fitting
}
