% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msnmm_pairs_methods.R
\name{summary.msnmm_pairs_fit}
\alias{summary.msnmm_pairs_fit}
\title{Summarize msnmm_pairs_fit}
\usage{
\method{summary}{msnmm_pairs_fit}(object, ...)
}
\description{
Produces a matrix with Estimate, SE, z, and p-values.
}
