% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msnmm_pairs_methods.R
\name{vcov.msnmm_pairs_fit}
\alias{vcov.msnmm_pairs_fit}
\title{Variance-covariance for msnmm_pairs_fit}
\usage{
\method{vcov}{msnmm_pairs_fit}(object, ...)
}
\description{
Variance-covariance for msnmm_pairs_fit
}
