% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R, R/builders_ml.R
\name{build_treatment_formulas_min}
\alias{build_treatment_formulas_min}
\title{Build treatment nuisance formulas (minimal)}
\usage{
build_treatment_formulas_min(
  K,
  stems,
  rhs_builder = rhs_linear,
  include_pastA = TRUE,
  ...
)

build_treatment_formulas_min(
  K,
  stems,
  rhs_builder = rhs_linear,
  include_pastA = TRUE,
  ...
)
}
\arguments{
\item{include_pastA}{logical; keep/remove past A terms (if builder adds them)}
}
\description{
Build treatment nuisance formulas (minimal)
}
