% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{create_wide_format}
\alias{create_wide_format}
\title{Create wide-style lagged columns Y_t, A_t, and other time-varying vars}
\usage{
create_wide_format(...)
}
\arguments{
\item{time_vars}{character vector of columns to expand}

\item{max_timepoints}{numeric K}
}
\description{
Create wide-style lagged columns Y_t, A_t, and other time-varying vars
}
