% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance.R
\name{fit_glm}
\alias{fit_glm}
\title{Specify a GLM learner (nuisance)}
\usage{
fit_glm(formula = NULL, family = gaussian(), vars = NULL)
}
\arguments{
\item{formula}{optional; if NULL, we build from vars}

\item{family}{glm family}

\item{vars}{character vector of column names}
}
\description{
Specify a GLM learner (nuisance)
}
