% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R, R/nuisance.R
\name{fit_glm}
\alias{fit_glm}
\title{Fit a GLM (wrapper)}
\usage{
fit_glm(formula = NULL, family = gaussian(), vars = NULL)

fit_glm(formula = NULL, family = gaussian(), vars = NULL)
}
\arguments{
\item{formula}{optional; if NULL, we build from vars}

\item{family}{glm family}

\item{vars}{character vector of column names}
}
\description{
Fit a GLM (wrapper)

Specify a GLM learner (nuisance)
}
