% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance.R
\name{fit_superlearner}
\alias{fit_superlearner}
\title{Specify a SuperLearner learner (nuisance)}
\usage{
fit_superlearner(SL.library, family = gaussian(), vars = NULL)
}
\description{
Specify a SuperLearner learner (nuisance)
}
