% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\name{fit_superlearner}
\alias{fit_superlearner}
\title{Fit with SuperLearner (generic)}
\usage{
fit_superlearner(
  formula,
  data,
  family = gaussian(),
  learners = c("SL.glm", "SL.ranger"),
  ...
)
}
\arguments{
\item{SL.library}{character vector (e.g. c("SL.ranger","SL.xgboost"))}
}
\description{
Fit with SuperLearner (generic)
}
