% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R, R/nuisance.R
\name{fit_superlearner}
\alias{fit_superlearner}
\title{Fit with SuperLearner (generic)}
\usage{
fit_superlearner(SL.library, family = gaussian(), vars = NULL)

fit_superlearner(SL.library, family = gaussian(), vars = NULL)
}
\arguments{
\item{SL.library}{character vector (e.g. c("SL.ranger","SL.xgboost"))}
}
\description{
Fit with SuperLearner (generic)

Specify a SuperLearner learner (nuisance)
}
