% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R, R/nuisance.R
\name{fit_xgboost}
\alias{fit_xgboost}
\title{Fit xgboost (simple recipe)}
\usage{
fit_xgboost(params = list(), nrounds = 200, vars = NULL)

fit_xgboost(params = list(), nrounds = 200, vars = NULL)
}
\arguments{
\item{params}{list passed to xgboost::xgb.train}

\item{nrounds}{number of boosting rounds}

\item{vars}{character vector of column names}
}
\description{
Fit xgboost (simple recipe)

Specify an xgboost learner (nuisance)
}
