% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance.R
\name{fit_xgboost}
\alias{fit_xgboost}
\title{Specify an xgboost learner (nuisance)}
\usage{
fit_xgboost(...)
}
\arguments{
\item{params}{list passed to xgboost::xgb.train}

\item{nrounds}{number of boosting rounds}

\item{vars}{character vector of column names}
}
\description{
Specify an xgboost learner (nuisance)
}
