% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{make_diffs}
\alias{make_diffs}
\title{Make diffs ΔY_{t+1} = Y_{t+1} - Y_t for t=0..K-1 (wide names)}
\usage{
make_diffs(...)
}
\description{
Make diffs ΔY_{t+1} = Y_{t+1} - Y_t for t=0..K-1 (wide names)
}
