% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{make_pastA}
\alias{make_pastA}
\title{Create past-treatment indicators (for "initiation" setups)}
\usage{
make_pastA(data, exposure, K)
}
\description{
Create past-treatment indicators (for "initiation" setups)
}
