% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msnmm_tv_pairs_min.R
\name{msnmm_tv_pairs_min}
\alias{msnmm_tv_pairs_min}
\title{Unified (m,k) SNMM g-estimation (minimal)}
\usage{
msnmm_tv_pairs_min(
  data,
  id,
  time = "yearms",
  ntimes,
  exposure = "A",
  outcome = "Y",
  time_vars_to_widen = c("A", "Y"),
  wide_ready = FALSE,
  outcome_nuisance_formulas,
  treatment_nuisance_formulas,
  blips,
  initiation = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{data}{wide data}

\item{id}{ignored when wide_ready=TRUE}

\item{ntimes}{integer horizon}

\item{outcome_nuisance_formulas}{list[\link{i}][\link{k-i+1}] of formulas}

\item{treatment_nuisance_formulas}{list[\link{i}] of formulas}

\item{blips}{list of blocks: list(m=,k=,preds=)}

\item{initiation}{logical}
}
\value{
object of class \code{msnmm_pairs_fit}
}
\description{
Unified (m,k) SNMM g-estimation (minimal)
}
